--- test-app-2/src/App.test.tsx
@@ -1,12 +1,0 @@
-import * as React from 'react';
-import { render } from '@testing-library/react';
-import { expect } from 'chai';
-import App from './App';
-
-describe('<App>', () => {
-  it('renders learn react link', () => {
-    const { getByText } = render(<App />);
-    const linkElement = getByText(/learn react/i);
-    expect(document.body.contains(linkElement));
-  });
-});
