--- test-app/.env
@@ -1,2 +1,0 @@
-SKIP_PREFLIGHT_CHECK=true
-FAST_REFRESH=true
