+++ test-app/src/App.test.tsx
@@ -1,9 +1,12 @@
-import React from 'react';
-import { render, screen } from '@testing-library/react';
+import * as React from 'react';
+import { render } from '@testing-library/react';
+import { expect } from 'chai';
 import App from './App';
 
-test('renders learn react link', () => {
-  render(<App />);
-  const linkElement = screen.getByText(/learn react/i);
-  expect(linkElement).toBeInTheDocument();
+describe('<App>', () => {
+  it('renders learn react link', () => {
+    const { getByText } = render(<App />);
+    const linkElement = getByText(/learn react/i);
+    expect(document.body.contains(linkElement));
+  });
 });
