+++ test-app/src/App.tsx
@@ -1,9 +1,9 @@
-import React from "react";
-import logo from "./logo.svg";
-import "./App.css";
+import React from 'react';
+import logo from './logo.svg';
+import './App.css';
 
-import { useDetectAdBlock } from "adblock-detect-react";
-import { detect } from "detect-browser";
+import { useDetectAdBlock } from 'adblock-detect-react';
+import { detect } from 'detect-browser';
 
 function App() {
   const adBlockDetected = useDetectAdBlock();
@@ -16,7 +16,7 @@ function App() {
         {adBlockDetected && (
           <>
             <img src={logo} className="App-logo" alt="logo" />
-            <div>{"An ad-blocker has been detected!"}</div>
+            <div>{'An ad-blocker has been detected!'}</div>
             <a
               className="App-link"
               href="https://www.npmjs.com/package/adblock-detect-react"
