+++ test-app/types/static.d.ts
